% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Get the feature importance of an \code{\link[=Xy]{Xy()}} simulation}
\usage{
importance(object, plot = TRUE)
}
\arguments{
\item{object}{an object of class \code{xy_sim}}

\item{plot}{a boolean signaling whether to print an importance plot}
}
\value{
Returns the effect importance of a cooked simulation recipe. The \code{plot}
argument can be used to create a feature importance plot.
}
\description{
Get the feature importance of an \code{\link[=Xy]{Xy()}} simulation
}
\details{
The importance is calculated as the relative local contribution of the effect,
for instance the absolute share of a linear simulated effect (x_1) to
the overall magnitude of y. The relative shares are then aggregated
over the whole simulated data set
}
\examples{
# create a simulation
linear_sim <- Xy() \%>\%
add_linear(p = 5) \%>\%
Xy::simulate(n=100)
# fetch the variable importance of the simualtion
linear_sim \%>\% importance()
}
