% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pulls}
\alias{pulls}
\alias{pull_x}
\alias{pull_y}
\alias{pull_e}
\alias{pull_xy}
\alias{pull_xye}
\title{Extract the simulated effects out of a cooked simulation recipe}
\usage{
pull_x(object)

pull_y(object)

pull_e(object)

pull_xy(object)

pull_xye(object)
}
\arguments{
\item{object}{an object of class \code{\link[=simulate]{xy_sim}}}
}
\value{
Returns the desired effects in a tibble.
}
\description{
Extract the simulated effects out of a cooked simulation recipe
}
\details{
The pull functions enable a convenient extraction of certain simulated
effects according to the desire of the user.
With \verb{pull_*} you can simply specify which effects you want to
extract out of the simulation
}
\examples{
# create a simulation
linear_sim <- Xy() \%>\%
  add_linear(p = 5) \%>\%
  simulate(n = 100)
# fetch the target
y <- linear_sim \%>\% pull_y()
# fetch all features
X <- linear_sim \%>\% pull_x()
# fetch the features and the target
xy <- linear_sim \%>\% pull_xy()
# fetch the residuals of the simulation
e <- linear_sim \%>\% pull_e()
}
