% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.Xy_sim.R
\name{varimp}
\alias{varimp}
\title{Variable Importance}
\usage{
varimp(object, use.noise = FALSE, plot = TRUE)
}
\arguments{
\item{object}{an object of class \code{Xy_sim}}

\item{use.noise}{a boolean indicating whether the noise of the process should
be added to the variable importance}

\item{plot}{a boolean specifying whether to print the variable importance}
}
\description{
Variable Importance
}
\examples{
# Visualize Feature Importance of a Simulation
my_simulation <- Xy()
varimp(my_simulation)
}
