% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xy.R
\name{Xy}
\alias{Xy}
\title{Xy}
\usage{
Xy(task = "regression", link = NULL, cutoff = NULL)
}
\arguments{
\item{task}{a character specifying the supervised learning task (e.g. "regression" or "classification").}

\item{link}{a link function to be used to transform the target.}

\item{cutoff}{a cutoff function. This function is applied after the link function.}
}
\value{
Returns an object of class \code{\link[=Xy]{xy_recipe}}. It is a simulation recipe.
which has to be cooked with \code{\link[=simulate]{simulate()}} to render a simulation.
}
\description{
The Xy function invokes a framework to simulate supervised learning tasks. With this function
you can specify the learning task (regression or classification) and you can
adjust specifics such as link functions and cutoff values (e.g. for classification).
The outcoming object serves as a mere simulation recipe and has to be enhanced
by functions which add effects to the simulation (see \code{\link{add_effects}}). Once
you have specified the recipe, you can cook the recipe with the \code{\link[=simulate]{simulate()}}
method.
}
\examples{
# create a recipe
recipe <- Xy() \%>\%
add_linear(p = 5, family = xy_normal())
}
\author{
Andre Bleier (\email{andre.bleier@statworx.com})
}
