% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xy_task.R
\name{Xy_task}
\alias{Xy_task}
\title{Xy Task}
\usage{
Xy_task(name = "regression", link = NULL, cutoff = NULL)
}
\arguments{
\item{name}{the name of the learning task}

\item{link}{a link function to be used to transform the target}

\item{cutoff}{a cutoff function. This function is applied after the link function.}
}
\value{
The function returns a list with a learning name and two transformation functions
}
\description{
This function creates a task list. Functions within this list are used to
transform the target variable of
}
\details{
The cutoff function is applied after transforming the target with the link function.
         It can be any function, however, in a classification learning task this function is used
         to classify the linktransformed target into zeroes and ones.
}
\examples{
# Regression
# In the regression case the link function is the identity link which
# means no transformation at all
my_regression_task <- Xy_task(name = "regression")
}
